refactor : add mappers, some names, and add custom errors

+++ src/main/java/ar/edu/unq/desapp/grupoG/backendapicryptoexchange/API/contracts/Transaction/UpdateTransactionRequest.java
@@ -11,5 +11,5 @@ public class UpdateTransactionRequest {
     @NotBlank
     @NotNull
     private String action;
-    private Integer user_id;
+    private Long user_id;
 }
