+++ src/main/java/ar/edu/unq/desapp/grupoG/backendapicryptoexchange/repositories/IUserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUserRepository extends CrudRepository<User, Integer> {
+public interface IUserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
